Extract AncestorBuilder
