+++ lib/rbs.rb
@@ -23,6 +23,7 @@ require "rbs/environment_loader"
 require "rbs/builtin_names"
 require "rbs/definition"
 require "rbs/definition_builder"
+require "rbs/definition_builder/ancestor_builder"
 require "rbs/variance_calculator"
 require "rbs/substitution"
 require "rbs/constant"
